Fixed bug in the add group
